requirements.txt
discord.py
pytube
ffmpy
colorlog
asyncio
cffi
pynacl
aiohttp
python-nmap
random-cat
pythonwhois
pillow
requests